+++ docs/news.rst
@@ -11,6 +11,7 @@ Psycopg 3.1 (unreleased)
 ------------------------
 
 - Add :ref:`Two-Phase Commit <two-phase-commit>` support (:ticket:`#72`).
+- Add `pq.PGconn.trace()` and related trace functions (:ticket:`#167`).
 
 
 Psycopg 3.0.6 (unreleased)